Add files via upload
We were having an issue where for some reason the fall data for volumetric water content was having a way higher result than it should. It has since been fixed and Amy helped clean up some other stuff in the volumetric water content. She recommended replacing geom_bar with stat_summary, so I did that. I believe it was the source of the issue?